Conflicts:
	modules/juce_audio_plugin_client/AAX/juce_AAX_Wrapper.cpp
	modules/juce_audio_plugin_client/AU/juce_AU_Wrapper.mm
	modules/juce_audio_plugin_client/RTAS/juce_RTAS_Wrapper.cpp
	modules/juce_audio_plugin_client/VST/juce_VST_Wrapper.cpp
	modules/juce_audio_processors/processors/juce_AudioProcessor.h

+++ modules/juce_core/native/juce_linux_Files.cpp
@@ -110,7 +110,7 @@ namespace
 {
     File juce_readlink (const String& file, const File& defaultFile)
     {
-        const int size = 8192;
+        const size_t size = 8192;
         HeapBlock<char> buffer;
         buffer.malloc (size + 4);
